feature-flag enhancements
